requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

#azure-functions
azure-functions==1.15.0
azure-storage-blob==12.17.0
pandas
io
